Update mongo_cmd.c
The mongo_cmd.c file is a series of MongoDB find and aggregation queries coded in C. It is possible to parameterize these queries to return specific fields or ranges. The aggregation queries return summaries of the data which may be preferred to avoid the transfer of large amounts of data across the network.